feat: refactor edit button
